fix: compute the meter output in double precision
_outModifier was declared float although ParamMTR_ChannelOutModifier
returns a double, so _counter * _outModifier was evaluated in single
precision. A float has a 24 bit mantissa, so a counter above 16.7
million was rounded: 50,000,001 went out as 50,000,000, and
UINT32_MAX became 4294967296.0f, which the conversion back to 32 bit
turned into 0.

The counter is a 32 bit register whose signedness is decided by
counterTypeSigned(), so the bit pattern has to survive the output
path untouched. It now does: a double represents every uint32_t
exactly, and KNXValue narrows through int64_t, which is modulo. With
a multiplier of 1 the register passes through unchanged, and each
datapoint reads it as configured. A signed counter on DPT 13 is no
longer off by up to 256 either.

+++ src/MeterChannel.h
@@ -18,7 +18,9 @@ class MeterChannel : public OpenKNX::Channel
     uint32_t _startTime = 0;
     uint32_t _lastTime = 0;
     uint8_t _mode = 0;
-    float _outModifier = 1.0;
+    // double, nicht float: float hat 24 Bit Mantisse und würde Zählerstände über
+    // 16,7 Mio. runden. Bei Multiplikator 1 muss ein uint32_t unverändert durchlaufen.
+    double _outModifier = 1.0;
     bool _locked = false;
     // bool _running = false;
     bool _afterStartup = false;
